+++ CHANGELOG.md
@@ -1,10 +1,38 @@
 # Change Log
 All notable changes to this project will be documented in this file.<br/>
 
+## [1.4] - 2016-12-24
+### Added
+- Output mirrors and their response times.</br>
+- More compatible *Arch Linux* versions.</br>
+
+### Changed
+- Minimized lines of code.</br>
+- Code optimizations.</br>
+- Minor visual improvements.</br>
+- Fixed issue while choosing \"Download and rank new mirrorlist (Country based)\".</br>
+
+## [1.3] - 2016-11-27
+### Added
+- Method to check how many mirrors are available for a mirrorlist.</br>
+- More compatible *Arch Linux* versions.</br>
+- Options to **backup** and **reset** mirrorlist.</br>
+- Checking functions on **steroids** e.g. *integer-check, exit-confirmation, refresh-confirmation*.</br>
+- Country based ranking after download.</br>
+
+### Changed
+- **Country based ranking** method for *new* and *existing* mirrorlist.</br>
+- In case of no internet connection, user can now backup or revert mirrolist.</br>
+- Method to check internet connection and tools' existance.</br>
+- Method to check bash version.</br>
+- Method to clean on exit.</br>
+- Code optimization.</br>
+- **Fixed:** internet checking function on slow connections.</br>
+
 ## [1.2] - 2016-11-16
 ### Added
 - Spinner on long running tasks to show progress.<br/>
-- More compatible Arch Linux versions.<br/>
+- More compatible *Arch Linux* versions.<br/>
 
 ### Changed
 - Minor code improvements.<br/>